move files around so maskLib can be installed
