+++ pyproject.toml
@@ -1,0 +1,11 @@
+[project]
+name = "maskLib"
+version = "0.1"
+description = "A package for making superconducting circuit designs"
+
+[build-system]
+requires = ["setuptools >= 61.0.0"]
+build-backend = "setuptools.build_meta"
+
+[tool.setuptools.packages.find]
+where = ["src"]
