+++ setup.py
@@ -1,0 +1,8 @@
+from setuptools import setup
+
+if __name__ == '__main__':
+    setup(
+        name='maskLib',
+        version='0.1',
+        packages=['maskLib'],
+    )
